+++ src/App.css
@@ -5,8 +5,3 @@
 body {
   font-family: Arial, sans-serif;
 }
-
-.article {
-  width: 960px;
-  margin: 0 auto;
-}
