+++ src/components/Article/Article.css
@@ -1,0 +1,4 @@
+.article {
+  width: 960px;
+  margin: 0 auto;
+}
